refactor(git): relocate platform/gitexec under platform/git
Move the shared git-exec package and its pinned-git test helper under a single platform/git umbrella: platform/gitexec becomes platform/git/exec, and platform/gitexec/gitexectest becomes platform/git/exectest — a sibling rather than nested, so the path no longer doubles "gitexec".

Package names are unchanged (gitexec, gitexectest), so this is an import-path move only: no call sites and no behavior change. It is the first step toward housing the change provider's bare-copy repo and auth plumbing alongside the shared git-exec package.
