+++ service/runway/server/BUILD.bazel
@@ -96,7 +96,7 @@ go_test(
     },
     deps = [
         "//api/base/mergestrategy/protopb:go_default_library",
-        "//platform/gitexec/gitexectest:go_default_library",
+        "//platform/git/exectest:go_default_library",
         "//runway/extension/merger/git:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
