+++ service/submitqueue/demo/requests/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
         "//platform/base/change/git:go_default_library",
         "//platform/base/change/github:go_default_library",
         "//platform/fakemarker:go_default_library",
-        "//platform/gitexec:go_default_library",
+        "//platform/git/exec:go_default_library",
         "//submitqueue/client:go_default_library",
         "@org_golang_x_sync//errgroup:go_default_library",
     ],
@@ -48,8 +48,8 @@ go_test(
     deps = [
         "//platform/base/change/git:go_default_library",
         "//platform/fakemarker:go_default_library",
-        "//platform/gitexec:go_default_library",
-        "//platform/gitexec/gitexectest:go_default_library",
+        "//platform/git/exec:go_default_library",
+        "//platform/git/exectest:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
     ],
