+++ test/e2e/submitqueue/BUILD.bazel
@@ -43,7 +43,7 @@ go_test(
         "//platform/extension/consumergate:go_default_library",
         "//platform/extension/consumergate/file:go_default_library",
         "//platform/extension/messagequeue/mysql:go_default_library",
-        "//platform/gitexec/gitexectest:go_default_library",
+        "//platform/git/exectest:go_default_library",
         "//platform/publish:go_default_library",
         "//submitqueue/core/batch:go_default_library",
         "//submitqueue/core/topickey:go_default_library",
